+++ skeletons/xer_decoder.c
@@ -296,7 +296,7 @@ xer_decode_general(const asn_codec_ctx_t *opt_codec_ctx,
 			 */
 			if(opt_unexpected_tag_decoder
 			&& opt_unexpected_tag_decoder(struct_key,
-					buf_ptr, ch_size) >= 0) {
+					buf_ptr, (size_t)ch_size) >= 0) {
 				/* Tag's processed fine */
 				ADVANCE(ch_size);
 				if(!ctx->phase) {
